Fibonnacci recursivo: C++, JS y WASM
- Fibonnacci supera con facilidad la capacidad de 32 bits con signo.
- Disponible ejecución en navegador.

+++ src/algorithm.cpp
@@ -1,4 +1,5 @@
-#include <emscripten.h>
+#include <cinttypes>
+#include <emscripten/emscripten.h>
 
 #ifdef __cplusplus
 #define EXTERN extern "C"
@@ -7,7 +8,7 @@
 #endif
 
 EXTERN EMSCRIPTEN_KEEPALIVE
-int fib(int n){
+uint64_t fib(uint64_t n){
   if (n <= 1) return n;
   return fib(n - 1) + fib(n - 2);
 }
